Export LKL syscall numbers, remove LKL-syscall-dependent functions from tests
LKL syscall numbers are needed in sgx-musl to replace the architecture ones.
We use a .c file that generates a .h file, so that build parameters (32/64 bits,
syscall compatibility tunings, etc.) are reflected in the generated header
automatically, and are not hardcoded from one LKL version to another.

Replace musl wrappers that use LKL syscalls in tests. E.g.: locatime() uses
do_tzset() which tries to __mmap_file("/etc/localtime"). We want these
tests to be entirely decoupled from LKL.